fix: resolve Tauri compilation errors and add UI styling
Compilation Fixes:
- Fix main.rs: Move cfg_attr before doc comments
- Fix security.rs: Use objc::msg_send! for NSWindow screenshot protection API
- Fix wallet.rs: Resolve type annotation errors in error handling
- Fix wallet.rs: Fix borrow-after-move issues with passphrase parameter
- Fix usb.rs: Add type casting for statvfs filesystem calculations

Dependencies:
- Add libc (0.2) for Unix filesystem operations
- Add cocoa (0.25) + objc (0.2) for macOS native APIs (screenshot protection)

UI Enhancements:
- Add comprehensive CSS styling to index.css (348 lines)
- Implement button styles: primary (blue), secondary (outlined), back button
- Add wallet card grid layout with hover effects
- Style form inputs with focus states and validation errors
- Add color scheme: blue (#3b82f6) primary, gray backgrounds, white cards
- Include responsive layout, shadows, transitions, and modern UI effects

Assets:
- Add placeholder app icons (32x32, 128x128, 256x256, .icns, .ico)

Build Status: ✅ Compilation successful with 15 warnings (unused imports)
Dev Server: ✅ Running on http://localhost:5173

+++ dashboard/src-tauri/src/commands/usb.rs
@@ -216,7 +216,7 @@ async fn get_available_space(path: &Path) -> Result<u64, String> {
         let mut stat: libc::statvfs = unsafe { std::mem::zeroed() };
 
         if unsafe { libc::statvfs(path_cstr.as_ptr(), &mut stat) } == 0 {
-            let available = stat.f_bavail * stat.f_bsize;
+            let available = (stat.f_bavail as u64) * (stat.f_bsize as u64);
             Ok(available)
         } else {
             Err("Failed to get filesystem stats".to_string())
